FEAT-#2606: Move the functions to another location; Add comments
